+++ lib/rubocop/cop/lint/nested_method_definition.rb
@@ -36,6 +36,13 @@ module RuboCop
       #       end
       #     end
       #   end
+      #
+      #   def foo
+      #     self.module_exec do
+      #       def bar
+      #       end
+      #     end
+      #   end
       class NestedMethodDefinition < Cop
         include OnMethodDef
         extend RuboCop::NodePattern::Macros
@@ -57,7 +64,7 @@ module RuboCop
               subject, = *child
               next if subject.lvar_type?
               yield child
-            elsif !ignored_child?(child)
+            elsif !scoping_method_call?(child)
               find_nested_defs(child, &block)
             end
           end
@@ -65,12 +72,17 @@ module RuboCop
 
         private
 
-        def ignored_child?(child)
-          eval_call?(child) || class_or_module_or_struct_new_call?(child)
+        def scoping_method_call?(child)
+          eval_call?(child) || exec_call?(child) ||
+            class_or_module_or_struct_new_call?(child)
         end
 
         def_node_matcher :eval_call?, <<-PATTERN
           (block (send _ {:instance_eval :class_eval :module_eval} ...) ...)
+        PATTERN
+
+        def_node_matcher :exec_call?, <<-PATTERN
+          (block (send _ {:instance_exec :class_exec :module_exec} ...) ...)
         PATTERN
 
         def_node_matcher :class_or_module_or_struct_new_call?, <<-PATTERN